homeassistant/components/frontend/www_static/polymer/bower_components/web-animations-js/test/blink/add-keyframes.html

<!doctype html>
<title>Web Animations API: Add Keyframe tests</title>
<script src="testharness/testharness.js"></script>
<script src="testharness/testharnessreport.js"></script>
<div id='log'></div>
<div id='div'></div>
<script>

test(function() {
  assert_throws({name: 'NotSupportedError'}, function() {
    div.animate([{height: '100px', composite: 'add'}, {height: '200px', composite: 'add'}], 1);
  }, 'Add animations should cause an exception.');
},
'Add animation tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#the-unaccumulated-animation-value-of-a-keyframe-animation-effect',
  assert: 'Only keyframes without add compositing considered',
  author: 'Shane Stephens'
});

test(function() {
  assert_throws({name: 'NotSupportedError'}, function() {
    div.animate([{height: '100px'}, {height: '200px', composite: 'add'}], 1);
  }, 'Animations with add keyframes should cause an exception.');
},
'Hybrid animation tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#the-unaccumulated-animation-value-of-a-keyframe-animation-effect',
  assert: 'Only keyframes without add compositing are considered',
  author: 'Shane Stephens'
});

</script>
